requirements.txt
acres==0.3.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.2.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
customtkinter==5.2.1
darkdetect==0.8.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.4
etelemetry==0.3.1
executing==2.1.0
filelock==3.16.1
fitz==0.0.1.dev2
greenlet==3.2.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
itemadapter==0.9.0
itemloaders==1.3.2
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
looseversion==1.3.0
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.1.3
packaging==24.1
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
pathlib==1.0.1
pbr==6.1.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
Protego==0.3.1
prov==2.0.1
psutil==6.1.0
pure_eval==0.2.3
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PyDispatcher==2.0.7
pydot==4.0.0
Pygments==2.18.0
PyMuPDF==1.24.14
pyOpenSSL==24.2.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyxnat==1.6.3
pyzmq==26.2.0
queuelib==1.7.0
RapidFuzz==3.10.1
rdflib==6.3.2
requests==2.32.3
requests-file==2.1.0
scipy==1.15.3
Scrapy==2.11.2
service-identity==24.2.0
setuptools==75.3.0
simplejson==3.20.1
six==1.16.0
SQLAlchemy==2.0.27
sqlparse==0.5.3
stack-data==0.6.3
stevedore==5.4.0
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
traits==7.0.2
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
virtualenvwrapper-win==1.2.7
w3lib==2.2.1
wcwidth==0.2.13
zope.interface==7.1.1
